Move more tests to Disk's own test module.
