+++ src/disk/tests.rs
@@ -2,6 +2,104 @@ use crate::chs::CHS;
 use crate::disk::Disk;
 use crate::partition::Partition;
 use crate::sector::Sector;
+use std::fs;
+use std::path::PathBuf;
+
+/// This should test disk geometry calculations but doesn't yet.
+#[test]
+fn disk_geometry() {
+    let mut my_disk = Disk::new("test.raw", 5000000);
+    my_disk
+        .partitions
+        .push(Partition::new(&my_disk, 1, 63, 4900000));
+    my_disk.write();
+}
+
+/// Test failure mode for creating a disk that is (much) too big.
+#[test]
+#[should_panic]
+fn disk_too_big() {
+    let mut my_disk = Disk::new("testdummy", 600000000000);
+    my_disk.partitions.push(Partition::new(&my_disk, 1, 63, 0));
+}
+
+/// Request an empty bootcode
+#[test]
+fn request_empty_bootcode() {
+    let bootcode: [u8; 446] = Disk::load_bootcode("EMPTY");
+    assert_eq!(bootcode, [0; 446]);
+}
+
+/// Request the MS-DOS 6.22 bootcode, compare length and a few bytes.
+#[test]
+fn request_msdos622_bootcode() {
+    let bootcode: [u8; 446] = Disk::load_bootcode("DOS622");
+    assert_eq!(bootcode.len(), 446);
+    assert_eq!(bootcode[0], 250);
+    assert_eq!(bootcode[217], 109);
+    assert_eq!(bootcode[218], 0);
+}
+
+/// Request a wrong type of bootcode
+#[test]
+#[should_panic]
+fn request_wrong_bootcode() {
+    Disk::load_bootcode("wrong_file.bin");
+}
+
+/// Create an empty disk. This will contain a few bytes that are non-zero.
+/// Read them back and compare to the expected result. This cleans up after itself.
+/// Uses an excessively unlikely filename so as not to clobber something already there.
+#[test]
+fn read_valid_sector() {
+    let mut my_disk = Disk::empty();
+    my_disk.path =
+        PathBuf::from("ff665c8ce7f5e1585ba2dcdc4109be56ef82dd0fccb5038449cc4fcf178345c1.raw");
+    my_disk.size = 50000000;
+    my_disk.write();
+    let mut reference: [u8; 512] = [
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+        0, 0,
+    ];
+    let mut reference_sector = Sector::new(0);
+    for (index, byte) in reference.iter().enumerate() {
+        reference_sector.write_byte(index, *byte);
+    }
+    let null_sector = my_disk.read_sector(0);
+    fs::remove_file("ff665c8ce7f5e1585ba2dcdc4109be56ef82dd0fccb5038449cc4fcf178345c1.raw")
+        .unwrap();
+    assert_eq!(null_sector, reference_sector);
+}
+
+/// Test for failure when trying to read a sector beyond the edge of a Disk.
+#[test]
+#[should_panic]
+fn read_sector_out_of_bounds() {
+    let my_disk = Disk::new(
+        "ff665c8ce7f5e1585ba2dcdc4109be56ef82dd0fccb5038449cc4fcf178345c1.raw",
+        50000000,
+    );
+    my_disk.write();
+    let reference = Sector::new(0);
+    let bad_sector = Disk::read_sector(&my_disk, 2000000000);
+    assert_eq!(bad_sector, reference);
+}
 
 /// Test building a Sector struct for the bootsector. Compare the default MS-DOS 6.22 boot sector to the
 /// statically provided byte-array in this function
